api/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
transformers
sentence-transformers
pinecone
supabase
requests
numpy
langfuse